// src/usecase/customer/update/update.customer.dto.ts
export interface InputUpdateCustomerDTO {
    id: string;
    name: string;
    address: {
        street: string;
        number: number;
        zip: string;
        city: string;
    };
}

export interface OutputUpdateCustomerDTO {
    id: string;
    name: string;
    address: {
        street: string;
        number: number;
        zip: string;
        city: string;
    };
}